Load OpenSSL through `autoload` consistently
We're still running into the `jruby` bug where the `OpenSSL` constant
can't be found even if it should be there. My hypotheses is that this is
happening because `net/http` still sets up an `autoload` for it. Since we
don't control that code, we can't really remove that `autoload`, so I'm
guessing doing the reserve (only `autoload` it and never explicitly
require it) should do the trick as well?